+++ spec/visual_grid/abort_test_for_render_fail_spec.rb
@@ -8,14 +8,14 @@ RSpec.describe 'Render Fail' do
   let(:eyes) { Applitools::Selenium::Eyes.new(runner: runner) }
   before(:all) do
     OpenSSL::SSL::VERIFY_PEER = OpenSSL::SSL::VERIFY_NONE
-    Applitools::EyesLogger.log_handler = Logger.new(STDOUT)
+    Applitools::EyesLogger.log_handler = Logger.new(STDOUT) unless ENV['TRAVIS']
   end
   it 'Failed render starts a session' do
     driver = eyes.open(
       app_name: 'Session for Render Fail',
       test_name: 'failed render',
       driver: web_driver,
-      viewport_size: { width: 1024, height: 768 }
+      viewport_size: { width: 800, height: 600 }
     )
     driver.get('http://applitools.com')
     eyes.check('step 1', Applitools::Selenium::Target.window.before_render_screenshot_hook('fail me!'))