+++ spec/visual_grid/browser_types_spec.rb
@@ -9,7 +9,7 @@ RSpec.shared_examples 'BrowserType' do |const, value|
     expect(subject.const_get(const).to_s).to eq(value)
   end
 end
-RSpec.describe BrowserTypes do
+RSpec.describe BrowserType do
   it_should_behave_like 'BrowserType', 'CHROME', 'chrome-0'
   it_should_behave_like 'BrowserType', 'CHROME_ONE_VERSION_BACK', 'chrome-1'
   it_should_behave_like 'BrowserType', 'CHROME_TWO_VERSIONS_BACK', 'chrome-2'
@@ -24,5 +24,6 @@ RSpec.describe BrowserTypes do
 
   it_should_behave_like 'BrowserType', 'IE_11', 'ie'
   it_should_behave_like 'BrowserType', 'IE_10', 'ie10'
-  it_should_behave_like 'BrowserType', 'EDGE', 'edge'
+  it_should_behave_like 'BrowserType', 'EDGE', 'edgelegacy'
+  it_should_behave_like 'BrowserType', 'EDGE_LEGACY', 'edgelegacy'
 end